Save point (tests don't pass because I discovered a bunch of unused calls in the Diffs Store)

+++ kernel/src/main/scala/net/lshift/diffa/kernel/differencing/DifferenceAggregationCache.scala
@@ -140,7 +140,7 @@ case class SequenceCacheKey(pair:DiffaPairRef, start:DateTime, end:DateTime)
  * Value stored in the sequence cache, representing the max sequence id found in the region along with the total
  * count of differences.
  */
-case class SequenceCacheValue(maxSeqId:Int, totalCount:Int) {
+case class SequenceCacheValue(maxSeqId:Long, totalCount:Int) {
   def toValidationKey = maxSeqId + "." + totalCount
 }
 object SequenceCacheValue {
